requirements.txt
streamlit

# Third-party components
streamlit-folium
streamlit_extras
streamlit-pandas-profiling
streamlit-aggrid
streamlit-image-coordinates
streamlit-plotly-events
streamlit-tags
streamlit-drawable-canvas

# etc
openai
folium
python-dotenv
